requirements.txt
# Requirements for SecureMessenger Pro
# Install with: pip install -r requirements.txt

# Image processing for steganography
Pillow>=9.0.0

# Cryptography library (AES encryption)
pycryptodome>=3.15.0

# For creating executable (optional)
pyinstaller>=6.0.0
